+++ examples/futures/Trading/CancelPlanOrder.php
@@ -11,7 +11,7 @@ $APIContract = new APIContractTrading(new CloudConfig([
     'memo' => "<your_memo>",
 ]));
 
-$response = $APIContract->cancelOrder("BTCUSDT", "<order_id>")['response'];
+$response = $APIContract->cancelPlanOrder("BTCUSDT", "234234123131231")['response'];
 
 
 echo json_encode($response);