+++ examples/futures/Trading/GetAssets.php
@@ -11,7 +11,5 @@ $APIContract = new APIContractTrading(new CloudConfig([
     'memo' => "<your_memo>",
 ]));
 
-$response = $APIContract->cancelOrder("BTCUSDT", "<order_id>")['response'];
-
-
+$response = $APIContract->getContractAssets()['response'];
 echo json_encode($response);