+++ examples/futures/Trading/SubmitLeverage.php
@@ -11,7 +11,22 @@ $APIContract = new APIContractTrading(new CloudConfig([
     'memo' => "<your_memo>",
 ]));
 
-$response = $APIContract->cancelOrder("BTCUSDT", "<order_id>")['response'];
+$response = $APIContract->submitLeverage(
+    "ETHUSDT",
+    "cross",
+    null,
+)['response'];
+
+
+echo json_encode($response);
+
+
+
+$response = $APIContract->submitLeverage(
+    "ETHUSDT",
+    "cross",
+    "5",
+)['response'];
 
 
 echo json_encode($response);